+++ src/main/scala/org/fs/mael/backend/http/HttpSettings.scala
@@ -2,6 +2,8 @@ package org.fs.mael.backend.http
 
 import org.eclipse.jface.preference.FieldEditorPreferencePage
 import org.eclipse.jface.preference.StringFieldEditor
+import org.fs.mael.backend.http.ui.CookiesConfigSetting
+import org.fs.mael.backend.http.ui.CookiesFieldEditor
 import org.fs.mael.ui.config.EmptyPreferencePage
 import org.fs.mael.ui.config.MFieldEditorPreferencePage
 import org.fs.mael.ui.config.MPreferencePageDescriptor
@@ -18,14 +20,14 @@ object HttpSettings {
 
     val pageDescriptors: Seq[MPreferencePageDescriptor[_ <: MFieldEditorPreferencePage]] = Seq(
       rootPageDescriptor,
-      MPreferencePageDescriptor("Headers", Some(rootPageDescriptor.name), classOf[HeadersPage])
+      MPreferencePageDescriptor("Headers", Some(rootPageDescriptor.name), classOf[GlobalHeadersPage])
     )
   }
 
   /** Setting pages for single download */
   object Local {
     val pageDescriptors: Seq[MPreferencePageDescriptor[_ <: MFieldEditorPreferencePage]] = Seq(
-      MPreferencePageDescriptor("Headers", None, classOf[HeadersPage])
+      MPreferencePageDescriptor("Headers", None, classOf[LocalHeadersPage])
     )
   }
 
@@ -36,14 +38,28 @@ object HttpSettings {
   val UserAgent: ConfigSetting[Option[String]] =
     ConfigSetting(prefix + ".userAgent", None)
 
+  val Cookies: ConfigSetting[Map[String, String]] =
+    new CookiesConfigSetting(prefix + ".cookies")
+
   //
   // Pages
   //
 
-  private class HeadersPage extends MFieldEditorPreferencePage(FieldEditorPreferencePage.FLAT) {
+  private class GlobalHeadersPage extends MFieldEditorPreferencePage(FieldEditorPreferencePage.FLAT) {
     override def createFieldEditors(): Unit = {
       row(UserAgent) { (setting, parent) =>
         new StringFieldEditor(setting.id, "User-Agent:", parent)
+      }
+    }
+  }
+
+  private class LocalHeadersPage extends MFieldEditorPreferencePage(FieldEditorPreferencePage.FLAT) {
+    override def createFieldEditors(): Unit = {
+      row(UserAgent) { (setting, parent) =>
+        new StringFieldEditor(setting.id, "User-Agent:", parent)
+      }
+      row(Cookies){ (setting, parent) =>
+        new CookiesFieldEditor(setting.id, "Cookies:", parent)
       }
     }
   }