+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * Added tray icon and minimize to tray setting
 * Added download speed and remaining time columns, implemented related tracking and calculations
 * Implemented HTTP header overrides
-* (Internal) Implemented backend-specific settings framework (unused yet)
+* (Internal) Implemented backend-specific settings framework (only used for HTTP headers so far)
   * (Internal) Implemented global defaults serving as template for local per-download settings
 * (Internal) Implemented migrations framework